fix(ios): more lenient parsing with incomplete stored configurations
A change related to Custom DNS supposed to be macOS only was also
affecting iOS stored configurations and could cause issues while
connecting when you upgrade the app *and* you don't launch it, but try
to connect anyway, from example from the iOS Settings app.

Configuration within Keychain would not be parsed, because it would have
missed the `dnsServers` property, causing the WG extension to not be
able to start (and connect).

Launching the app and waiting for it to send a new configuration with a
default value for the `dnsServers` property was sufficient to workaround
the situation; but this fix make sure that we cover extensions upgrades
without main app launches.

Jira-Id: VPNAPPL-2840